add package-info.java to rest directory, move sample programs from main of classes to their own files under samples/ava/functional
