+++ samples/java/functional/src/rest/package-info.java
@@ -1,0 +1,11 @@
+/*
+# Licensed Materials - Property of IBM
+# Copyright IBM Corp. 2017 
+ */
+
+/**
+ * Samples using Java REST api wrappers
+ */
+
+package rest;
+
